home made atmospheric density
